RAIN-12333
Here are my suggestions for improvement to the top-level page. 
I changed this sentence "The source for the officially supported Lacework agent releases." because it implies that we are providing the source code to the agents. Plus the fact that this is a GitHub site implies that it is source, also.